Formatting

+++ src/csvreader.cpp
@@ -134,7 +134,7 @@ namespace ExtCsvLoader
 			if (m_with_row_header)
 			{
 #pragma omp parallel for
-				for (std::int64_t column_index = 0; column_index < m_nrOfColumns; ++column_index)
+				for (std::ptrdiff_t column_index = 0; column_index < m_nrOfColumns; ++column_index)
 				{
 					header.getAs(column_index + 1, m_column_header[column_index]);
 				}
@@ -142,7 +142,7 @@ namespace ExtCsvLoader
 			else
 			{
 #pragma omp parallel for
-				for (std::int64_t column_index = 0; column_index < m_nrOfColumns; ++column_index)
+				for (std::ptrdiff_t column_index = 0; column_index < m_nrOfColumns; ++column_index)
 				{
 					header.getAs(column_index, m_column_header[column_index]);
 				}
